requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
autopep8==1.7.0
backports.zoneinfo;python_version<"3.9"
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
dj-static==0.0.6
Django==4.0.7
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-timezone-field==5.0
djangorestframework==3.13.1
idna==3.3
kombu==5.2.4
mysqlclient==2.1.1
packaging==21.3
prompt-toolkit==3.0.30
pycodestyle==2.9.1
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
toml==0.10.2
tzdata==2022.2
urllib3==1.26.12
vine==5.0.0
waitress==2.1.2
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
